scripts/commit_warnings_extractor/requirements.txt
git-python
tqdm
